major change to manuscript_figs_tables.ipynb
Now requies the directory containing manuscript files is a sister to this directory.

Generates the entire supplementary file, only tables/figs actually referenced
in the main manuscript files are written. Produces supp_material_generated.tex
The content of that needs to be copied to the appropriate file on Sharelatex!

Many changes to the formatting, numerous fixes (cogent3 tables to be used
to produce latex MUST have their caption as the title, NOT legend)